Solved conflicts. Implemented movements. Moved prints for debugging to a separate file.
