+++ scenes/basic_scene.rt
@@ -4,5 +4,3 @@ L 0,50,30                 0.9          255,255,255
 
 pl 0,0,0          0,1,0                 255,0,0
 sp 0,5,-10                  5           0,255,0
-sp 2,5,-10                  5           255,255,0
-sp -2,5,-10                  5           0,255,255
