+++ src/parser/add_shapes.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   add_shapes.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: henbuska <henbuska@student.hive.fi>        +#+  +:+       +#+        */
+/*   By: irychkov <irychkov@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/20 11:01:33 by henbuska          #+#    #+#             */
-/*   Updated: 2025/02/20 16:25:50 by henbuska         ###   ########.fr       */
+/*   Updated: 2025/02/20 22:35:47 by irychkov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,10 +22,13 @@ void	add_sphere(t_rt *rt, char **coordinates, char **colors, double diameter)
 	rt->scene->shapes[i].radius = diameter / 2;
 	rt->scene->shapes[i].material.color = string_to_color(colors);
 	rt->scene->shapes[i].material.ambient = rt->scene->ambient.ratio;
-	rt->scene->shapes[i].material.diffuse = 0.8;
-	rt->scene->shapes[i].material.specular = 0.4;
-	rt->scene->shapes[i].material.shininess = 150.0;
-	rt->scene->shapes[i].transform = identity_matrix(4);
+	rt->scene->shapes[i].material.diffuse = 0.3;
+	rt->scene->shapes[i].material.specular = 1.0;
+	rt->scene->shapes[i].material.shininess = 300.0;
+	/* rt->scene->shapes[i].material.reflective = 0.9;
+	rt->scene->shapes[i].material.transparency = 0.0;
+	rt->scene->shapes[i].material.refractive_index = 1.0; */
+	rt->scene->shapes[i].transform = identity_matrix();
 	rt->scene->shape_count++;
 }
 
@@ -41,7 +44,10 @@ void	add_plane(t_rt *rt, char **coordinates, char **normal, char **colors)
 	rt->scene->shapes[i].material.diffuse = 0.8;
 	rt->scene->shapes[i].material.specular = 0.4;
 	rt->scene->shapes[i].material.shininess = 150.0;
-	rt->scene->shapes[i].transform = identity_matrix(4);
+	/* rt->scene->shapes[i].material.reflective = 0.2;
+	rt->scene->shapes[i].material.transparency = 0.0;
+	rt->scene->shapes[i].material.refractive_index = 0.0; */
+	rt->scene->shapes[i].transform = identity_matrix();
 	rt->scene->shape_count++;
 }
 
@@ -57,6 +63,6 @@ void	add_cylinder(t_rt *rt, char **coordinates, char **normal, char **colors)
 	rt->scene->shapes[i].material.diffuse = 0.8;
 	rt->scene->shapes[i].material.specular = 0.4;
 	rt->scene->shapes[i].material.shininess = 150.0;
-	rt->scene->shapes[i].transform = identity_matrix(4);
+	rt->scene->shapes[i].transform = identity_matrix();
 	rt->scene->shape_count++;
 }
